load_image/pyproject.toml
[project]
name = "load_images"
version = "0.1.0"
description = "Loads images for p-hash"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "annotated-doc==0.0.4",
  "annotated-types==0.7.0",
  "anyio==4.12.0",
  "click==8.3.1",
  "fastapi==0.124.4",
  "h11==0.16.0",
  "idna==3.11",
  "psycopg2-binary==2.9.11",
  "pydantic==2.12.5",
  "pydantic_core==2.41.5",
  "starlette==0.50.0",
  "typing-inspection==0.4.2",
  "typing_extensions==4.15.0",
  "uvicorn==0.38.0",
]

[tool.setuptools.packages.find]
where = ["."]


